[dynamo] Fix side-effect handling for pre-existing `collections.deque`
Previously we never replayed side effects to `DequeVariable` with a
source; the bug was already in the `test_deque_input` test, but went
unnoticed because we didn't check the deque objects.

This patch adds limited but practical support for this (see comments in
`side_effects.py` for why limited), and updates the deque tests to check
for this.

ghstack-source-id: b060b2f257b0cf8cd1737664521ee8e770c8dca5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/141714