[dynamo] Validate `mutation_type` and `source` in `VariableTracker.__init__`
As title, this also uncovered a few invalid use cases; the cases that
cause error are fixed in separate patches prior to this patch, and the
rest are fixed in this patch.

This patch also moves a few `.source` mutation to variable construction,
to increase the coverage of the validation.

Fixes #133027.

ghstack-source-id: 17faa9aafd78750dd63ba66fbe3379b192112d5c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/141717